requirements.txt
# ── Core ───────────────────────────────────────────────
torch>=2.1.0
torchvision>=0.16.0
torchmetrics>=1.2.0

# ── Data / Images ─────────────────────────────────────
Pillow>=10.0
albumentations>=1.3.0
numpy>=1.24

# ── Config & CLI ──────────────────────────────────────
pyyaml>=6.0
typer[all]>=0.9.0

# ── Logging ───────────────────────────────────────────
rich>=13.0

# ── Testing ───────────────────────────────────────────
pytest>=7.4
pytest-cov>=4.1

# ── Utilities ─────────────────────────────────────────
tqdm>=4.66
matplotlib>=3.7